+++ src/viewModel/adminHook.js
@@ -1,8 +1,9 @@
-import { client } from './client';
+import { api } from '../utils/api';
+import { useState } from 'react';
 
 export const createSession = async (sessionName, date) => {
   try {
-    const response = await client.post('/session/createSession', {
+    const response = await api.post('/session/createSession', {
       name: sessionName,
       date: date
     });
@@ -13,9 +14,41 @@ export const createSession = async (sessionName, date) => {
     throw error;
   }
 };
+function useCreateSession(onSuccess) {
+  const [sessionName, setSessionName] = useState("");
+  const [date, setDate] = useState("");
+
+  const onChangeSessionName = (value) => {
+    setSessionName(value);
+  };
+
+  const onChangeDate = (value) => {
+    setDate(value);
+  };
+
+  const onPressCreateSession = async () => {
+    try {
+      const formattedDate = new Date(date).toISOString();
+      console.log("Formatted Date:", formattedDate);
+      await createSession(sessionName, formattedDate);
+      alert(`새로운 세션이 생성되었습니다.`);
+      onSuccess();
+    } catch (error) {
+      console.error("Failed to create session:", error);
+    }
+  };
+
+  return {
+    sessionName,
+    date,
+    onChangeSessionName,
+    onChangeDate,
+    onPressCreateSession,
+  };
+}
 export const getSessions = async () => {
     try {
-      const response = await client.get('/session/sessions');
+      const response = await api.get('/session/sessions');
   
       if (Array.isArray(response.data)) {
         console.log("sessions fetched:", response.data); 
@@ -33,7 +66,7 @@ export const getSessions = async () => {
   export const startAttendCheck = async (sessionId) => {
     try {
       console.log(`Starting attendance check for session ID: ${sessionId}`); 
-      const response = await client.post(`/session/startAttendCheck/${sessionId}`);
+      const response = await api.post(`/session/startAttendCheck/${sessionId}`);
       return response.data;
     } catch (error) {
       console.error('Error starting attendance check:', error);
@@ -43,7 +76,7 @@ export const getSessions = async () => {
   
   export const endAttendCheck = async () => {
     try {
-      const response = await client.delete('/session/endAttendCheck');
+      const response = await api.delete('/session/endAttendCheck');
       return response.data;
     } catch (error) {
       console.error('Error ending attendance check:', error);
@@ -51,8 +84,10 @@ export const getSessions = async () => {
     }
   };
   export const deleteSession = async (sessionId) => {
-    const response = await client.delete(`/session/deleteSession/${sessionId}`);
+    const response = await api.delete(`/session/deleteSession/${sessionId}`);
     console.log("session deleted", response);
     
     return response.data;
   };
+
+export { useCreateSession };
